Fix issues with rebase, defer /role reply
