+++ src/pg-storage-client.ts
@@ -17,6 +17,7 @@ export default class PGStorageClient {
         'tracking_channels': 'CREATE TABLE tracking_channels (guild_id BIGINT PRIMARY KEY, channel_id BIGINT);',
         'player_hiscore_status': 'CREATE TABLE player_hiscore_status (rsn VARCHAR(12) PRIMARY KEY, on_hiscores BOOLEAN);',
         'bot_counters': 'CREATE TABLE bot_counters (user_id BIGINT PRIMARY KEY, counter INTEGER);',
+        'privileged_roles': 'CREATE TABLE privileged_roles (guild_id BIGINT PRIMARY KEY, role_id BIGINT);',
         'misc_properties': 'CREATE TABLE misc_properties (name VARCHAR(32) PRIMARY KEY, value VARCHAR(2048));'
     };
 
@@ -185,6 +186,23 @@ export default class PGStorageClient {
     
     async writeBotCounter(userId: Snowflake, counter: number): Promise<void> {
         await this.client.query('INSERT INTO bot_counters VALUES ($1, $2) ON CONFLICT (user_id) DO UPDATE SET counter = EXCLUDED.counter;', [userId, counter]);
+    }
+
+    async fetchAllPrivilegedRoles(): Promise<Record<Snowflake, Snowflake>> {
+        const result: Record<Snowflake, Snowflake> = {};
+        const queryResult = await this.client.query<{guild_id: Snowflake, role_id: Snowflake}>('SELECT * FROM privileged_roles;');
+        for (const row of queryResult.rows) {
+            result[row.guild_id] = row.role_id;
+        }
+        return result;
+    }
+    
+    async deletePrivilegedRole(guildId: Snowflake, roleId: Snowflake): Promise<void> {
+        await this.client.query('DELETE FROM privileged_roles WHERE guild_id = $1 AND role_id = $2;', [guildId, roleId]);
+    }
+    
+    async writePrivilegedRole(guildId: Snowflake, roleId: Snowflake): Promise<void> {
+        await this.client.query('INSERT INTO privileged_roles VALUES ($1, $2) ON CONFLICT (guild_id) DO UPDATE SET role_id = EXCLUDED.role_id;', [guildId, roleId]);
     }
     
     async fetchMiscProperty(name: MiscPropertyName): Promise<string | null> {
